papers.tex: Move all page numbers to the same height
We won't be making use of the intended workaround of moving the page numbers to
cover those on submission originals; instead, I'll manually redact the current
page numbers with Acrobat prior to sending it for publication. In future years,
we should fix the main-matter.tex code to extend vboxes upward and cover old
numbering systems.